Add issue-modifying functionality to rietveld package
BUG=skia:4455
R=stephana@google.com

Review URL: https://codereview.chromium.org/1405753003 .

+++ go/autoroll/autoroll.go
@@ -25,7 +25,7 @@ var r = rietveld.New(RIETVELD_URL, nil)
 type AutoRollIssue struct {
 	Closed    bool
 	Committed bool
-	Issue     int
+	Issue     int64
 	Modified  time.Time
 }
